[IMP] base,*: support image/webp image format
*: mail, mrp, test_website, web, web_editor

Before this commit '.webp' images could not be used in odoo.

After this commit '.webp' images can be uploaded to odoo.
- can be used in image field
- can be used in HTML field image
- can be used in mails and website
- can be transformed (shape mask, filter effect, crop, rotate, resize)

Quality cannot be adjusted because this is would be achieved by a Pillow
plugin that we prefer not to use for security reasons because it relies
on the 'libwebp' native library.

task-2774352